Added new recordQRAttendance Router for conciseness. Prisma schema updated to include attendance token table
