+++ prisma/schema.prisma
@@ -1,4 +1,3 @@
-
 generator client {
   provider = "prisma-client-js"
 }
@@ -11,7 +10,7 @@ datasource db {
 
 model User {
   id               String    @id @default(cuid())
-  role             String  @db.VarChar(10)
+  role             String    @db.VarChar(10)
   name             String?
   email            String?   @unique
   emailVerified    DateTime?
@@ -52,7 +51,7 @@ model VerificationToken {
 model Lecture {
   id                String            @id @default(cuid())
   course            Course            @relation(fields: [courseId], references: [id])
-  courseId          String            
+  courseId          String
   lectureDate       DateTime
   attendanceEntries AttendanceEntry[]
 }
@@ -72,12 +71,13 @@ model Course {
   courseCode  String         @unique @db.VarChar(255)
   name        String         @db.VarChar(255)
   lmsType     String
-  lmsId       String?        
+  lmsId       String?
   dateCreated DateTime       @default(now())
   StartDate   DateTime?
   EndDate     DateTime?
   lectures    Lecture[]
   enrollments CourseMember[]
+
   @@unique([lmsType, lmsId])
 }
 
@@ -89,9 +89,10 @@ model CourseMember {
   courseId          String
   course            Course            @relation(fields: [courseId], references: [id])
   //user              User              @relation(fields: [email], references: [email]) // For now we don't need to focus on this 
-  dateEnrolled    DateTime           @default(now())
+  dateEnrolled      DateTime          @default(now())
   role              String            @default("student")
   attendanceEntries AttendanceEntry[]
+
   @@unique([email, courseId])
 }
 
@@ -102,4 +103,14 @@ model qrcode {
   code      String
   createdAt DateTime @default(now())
   expiresAt DateTime
+}
+
+model AttendanceToken {
+  id        String   @id @default(cuid())
+  token     String
+  lectureId String
+  createdAt DateTime @default(now())
+
+  // We could implement a system where these are auto deleted after expiration (e.g 10 mins)
+  // expiresAt DateTime 
 }
