+++ server/routes/qr.ts
@@ -7,11 +7,6 @@ export const zCreateQRCode = z.object({
   courseId: z.string()
 });
 
-export const zValidateCode = z.object({
-  qr: z.string(),
-  courseId: z.string()
-});
-
 export const qrRouter = router({
   CreateNewQRCode: publicProcedure
     .input(zCreateQRCode)
@@ -54,51 +49,4 @@ export const qrRouter = router({
         throw new Error('Error Removing QR code');
       }
     }),
-
-  //Search qrcode table for QRCode sent by the user
-
-  ValidateQRCode: publicProcedure.input(zValidateCode).query(async ({ input }) => {
-    try {
-      const qr = input.qr;
-      const courseId = input.courseId;
-
-      const storedQRCode = await prisma.qrcode.findFirst({
-        where: {
-          code: qr,
-          courseId: courseId
-        }
-      });
-
-      //Check if we find a code in our DB that matches the code sent by the user
-      if (!!storedQRCode) {
-        return { success: true };
-      } else {
-        return { success: false };
-      }
-    } catch (error) {
-      throw new Error('Error finding QR code');
-    }
-  }),
-
-  CreateAttendanceToken: publicProcedure
-    .input(zValidateCode)
-    .query(async ({ input }) => {
-      try {
-        console.log(input);
-        const storedQRCode = await prisma.qrcode.findFirst({
-          where: {
-            code: input.qr
-          }
-        });
-
-        //Check if we find a code in our DB that matches the code sent by the user
-        if (!!storedQRCode) {
-          return { success: true };
-        } else {
-          return { success: false };
-        }
-      } catch (error) {
-        throw new Error('Error finding QR code');
-      }
-    })
 });
